Moves demo related code into separate ‘demo’ package.
